Add execute command service
