+++ homeassistant/components/system_bridge/icons.json
@@ -56,6 +56,9 @@
     }
   },
   "services": {
+    "execute_command": {
+      "service": "mdi:console-line"
+    },
     "get_process_by_id": {
       "service": "mdi:console"
     },
