+++ homeassistant/components/system_bridge/services.yaml
@@ -89,3 +89,23 @@ power_command:
             - "restart"
             - "shutdown"
             - "sleep"
+execute_command:
+  fields:
+    bridge:
+      required: true
+      selector:
+        device:
+          integration: system_bridge
+    command_id:
+      required: true
+      selector:
+        text:
+    timeout:
+      required: false
+      default: 300
+      example: 300
+      selector:
+        number:
+          min: 1
+          max: 600
+          unit_of_measurement: seconds
